pyproject.toml
[tool.poetry]
name = "python-basic-diploma"
version = "0.1.0"
description = ""
authors = ["DenisKhudyakov <denis.hudyakov27@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.1"
aiogram = "^3.7.0"
psycopg2-binary = "^2.9.9"
requests = "^2.32.3"
redis = "^5.0.6"
celery = "^5.4.0"
pytelegrambotapi = "^4.19.1"


[tool.poetry.group.lint.dependencies]
black = "^24.4.2"
mypy = "^1.10.0"
isort = "^5.13.2"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
